+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [v3.6.2]
 
+### Changed
+
+- Make Syscheck case insensitive. ([#1349](https://github.com/wazuh/wazuh/pull/1349))
+
 ### Fixed
 
 - Fix 0 value at check options from Syscheck. ([1209](https://github.com/wazuh/wazuh/pull/1209))
@@ -14,7 +18,7 @@ All notable changes to this project will be documented in this file.
 - Fix stack overflow when monitoring deep files. ([#1239](https://github.com/wazuh/wazuh/pull/1239))
 - Fix bug when running quick commands with timeout of 1 second. ([#1259](https://github.com/wazuh/wazuh/pull/1259))
 - Fixed getting RAM memory information from mac OS X and FreeBSD agents. ([#1203](https://github.com/wazuh/wazuh/pull/1203))
-
+- Set connection timeout for Auth server ([#1336](https://github.com/wazuh/wazuh/pull/1336))
 
 ## [v3.6.1] 2018-09-07